Fix constant arithmetic overflow warning by using correct types

+++ ITKImageProcessingFilters/ITKDiscreteGaussianImage.h
@@ -32,7 +32,7 @@ class ITKImageProcessing_EXPORT ITKDiscreteGaussianImage : public ITKImageProces
   Q_OBJECT
   PYB11_CREATE_BINDINGS(ITKDiscreteGaussianImage SUPERCLASS ITKImageProcessingBase)
   PYB11_PROPERTY(FloatVec3Type Variance READ getVariance WRITE setVariance)
-  PYB11_PROPERTY(double MaximumKernelWidth READ getMaximumKernelWidth WRITE setMaximumKernelWidth)
+  PYB11_PROPERTY(int32_t MaximumKernelWidth READ getMaximumKernelWidth WRITE setMaximumKernelWidth)
   PYB11_PROPERTY(FloatVec3Type MaximumError READ getMaximumError WRITE setMaximumError)
   PYB11_PROPERTY(bool UseImageSpacing READ getUseImageSpacing WRITE setUseImageSpacing)
 
@@ -46,8 +46,8 @@ public:
   SIMPL_FILTER_PARAMETER(FloatVec3Type, Variance)
   Q_PROPERTY(FloatVec3Type Variance READ getVariance WRITE setVariance)
 
-  SIMPL_FILTER_PARAMETER(double, MaximumKernelWidth)
-  Q_PROPERTY(double MaximumKernelWidth READ getMaximumKernelWidth WRITE setMaximumKernelWidth)
+  SIMPL_FILTER_PARAMETER(int32_t, MaximumKernelWidth)
+  Q_PROPERTY(int32_t MaximumKernelWidth READ getMaximumKernelWidth WRITE setMaximumKernelWidth)
 
   SIMPL_FILTER_PARAMETER(FloatVec3Type, MaximumError)
   Q_PROPERTY(FloatVec3Type MaximumError READ getMaximumError WRITE setMaximumError)
